Create resume controller and write some tests.
