+++ controller/resume/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  upsert: require('./upsert'),
+  'update-theme': require('./update-theme')
+};
